moving recipe in repo, updating license file

+++ ci-support/conda_upload.sh
@@ -25,10 +25,19 @@ echo "Cloning recipes"
 git clone git://github.com/CDAT/conda-recipes
 cd conda-recipes
 # uvcdat creates issues for build -c uvcdat confises package and channel
-rm -rf uvcdat
 export BRANCH=${CIRCLE_BRANCH}
+if [[ -d uvcdat ]]; then
+    rm -rf uvcdat
+fi
+if [[ -d libdrs_c ]]; then
+    rm -rf libdrs_c
+fi
+if [[ -d libdrs_f ]]; then
+    rm -rf libdrs_f
+fi
+ln -s ../recipe/libdrs_c libdrs_c
+ln -s ../recipe/libdrs_f libdrs_f
 python ./prep_for_build.py  -b ${BRANCH}
-
 conda build ${PKG_NAME}_c -c conda-forge 
 conda build ${PKG_NAME}_f -c conda-forge 
 anaconda -t $CONDA_UPLOAD_TOKEN upload -u $USER -l $LABEL $CONDA_BLD_PATH/$OS/${PKG_NAME}-$VERSION.`date +%Y*`0.tar.bz2 --force
